# [PERF] product: Improve name_search performance