@robodoo 2023-11-02
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/140804).