@vava-odoo 2023-11-06
Hi @BHanq 
There is another fix on this: https://github.com/odoo/odoo/pull/139844 
With a really simpler solution, timings look similar... Isn't it sufficient?