@BHanq 2023-11-13
Hello @vava-odoo, 

The PR has been neglected for a long time, so I just added into my PR, in order to have more impact on this as it has the same goal.

The timing looks similar but shows totally different workflow, as the other PR only add a condition in order to avoid an useful calculation. While my whole PR stands for improving the code into the condition. As stated in @awa-odoo comment, `# FIXME awa: this is really not performant` a refactor is highly recommended. 

If you need a better explanation, tell me what I should explain broader. 

Thanks for your time