@BHanq 2023-11-24
targeting 15.0 on https://github.com/odoo/odoo/pull/143543